+++ scripts/generate-docs.ts
@@ -207,7 +207,7 @@ async function runHydrogenGenerator(args: Partial<Options> = {}) {
         'foundation/useServerState',
         'foundation/useShop',
         'foundation/useQuery',
-        'hooks/useShopQuery'
+        'hooks/useShopQuery',
       ],
     }),